feat(scoop-cache): Allow multiple apps to be passed (#141)

+++ libexec/scoop-cache.ps1
@@ -1,59 +1,55 @@
-# Usage: scoop cache [rm|show] [app] [options]
-# Summary: Show or clear the download cache
-# Help: Scoop caches downloads so you don't need to download the same files
-# when you uninstall and re-install the same version of an app.
+# Usage: scoop cache [<SUBCOMMAND>] [<OPTIONS>] [<APP>...]
+# Summary: Show or clear the download cache.
+# Help: Scoop caches downloaded files to remove the need for repeated downloads of same files.
 #
 # You can use
 #   scoop cache show
 # to see what's in the cache, and
-#   scoop cache rm <app>
+#   scoop cache rm git
 # to remove downloads for a specific app.
 #
-# To clear everything in your cache, use:
+# To clear everything in cache, use:
 #   scoop cache rm *
+#
+# Subcommands:
+#   rm              Remove an application specific files from cache.
+#   show            Show an overview of all cached files. Default command when any is provided.
 #
 # Options:
 #   -h, --help      Show help for this command.
 
-param($cmd, $app)
-
-. (Join-Path $PSScriptRoot '..\lib\help.ps1')
+'core', 'Cache', 'getopt', 'help' | ForEach-Object {
+    . (Join-Path $PSScriptRoot "..\lib\$_.ps1")
+}
 
 Reset-Alias
 
-function cacheinfo($file) {
-    $app, $version, $url = $file.name -split '#'
-    $size = filesize $file.length
-    return New-Object PSObject -prop @{ 'app' = $app; 'version' = $version; 'url' = $url; 'size' = $size }
-}
+$opt, $arguments, $err = getopt $args
+if ($err) { Stop-ScoopExecution -Message "scoop cache: $err" -ExitCode 2 }
 
-function show($app) {
-    $files = @(Get-ChildItem $SCOOP_CACHE_DIRECTORY | Where-Object -Property 'Name' -Match "^$app")
-    $total_length = ($files | Measure-Object length -Sum).sum -as [double]
-
-    $f_app = @{ 'Expression' = { "$($_.app) ($($_.version))" } }
-    $f_url = @{ 'Expression' = { $_.url }; 'Alignment' = 'Right' }
-    $f_size = @{ 'Expression' = { $_.size }; 'Alignment' = 'Right' }
-
-
-    $files | ForEach-Object { cacheinfo $_ } | Format-Table $f_size, $f_app, $f_url -AutoSize -HideTableHeaders
-
-    "Total: $($files.length) $(pluralize $files.length 'file' 'files'), $(filesize $total_length)"
-}
-
+$cmd = if ($arguments[0]) { $arguments[0] } else { 'show' }
+$isShow = $cmd -eq 'show'
+$applications = $arguments[1..($arguments.Count)]
 $exitCode = 0
-switch ($cmd) {
-    'rm' {
-        if (!$app) { Stop-ScoopExecution -Message 'Parameter <app> missing' -Usage (my_usage) }
-        Join-Path $SCOOP_CACHE_DIRECTORY "$app#*" | Remove-Item -Force -Recurse
-        Join-Path $SCOOP_CACHE_DIRECTORY "$app.txt" | Remove-Item -ErrorAction 'SilentlyContinue' -Force -Recurse
-    }
-    'show' {
-        show $app
-    }
-    default {
-        show
+$problems = 0
+
+if ($cmd -notin @('rm', 'show')) { Stop-ScoopExecution -Message "Unknown subcommand: '$cmd'" -Usage (my_usage) }
+if (!$isShow -and !$applications) { Stop-ScoopExecution -Message 'Parameter <APP> is required for ''rm'' subcommand' -Usage (my_usage) }
+
+if ($isShow) {
+    Show-CachedFileList -ApplicationFilter $applications
+} else {
+    foreach ($app in $applications) {
+        try {
+            Join-Path $SCOOP_CACHE_DIRECTORY "$app#*" | Remove-Item -ErrorAction 'Stop' -Force -Recurse
+            Join-Path $SCOOP_CACHE_DIRECTORY "$app.txt" | Remove-Item -ErrorAction 'SilentlyContinue' -Force -Recurse
+        } catch {
+            Write-UserMessage -Message "Removing ${app}: $($_.Exception.Message)" -Err
+            ++$problems
+        }
     }
 }
+
+if ($problems -gt 0) { $exitCode = 10 + $problems }
 
 exit $exitCode
